requirements.txt
flask
streamlit
scikit-learn
pandas
numpy
openai
python-dotenv
pillow
joblib